# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "YEAH API",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/data/jeniskelamin": {
            "get": {
                "tags": [
                    "Data"
                ],
                "summary": "ambil data jenis kelamin",
                "description": "ambil data jenis kelamin",
                "operationId": "jeniskelamin",
                "responses": {
                    "200": {
                        "description": "sukses ambil data jenis kelamin"
                    },
                    "400": {
                        "description": "gagal mendapatkan data jenis kelamin"
                    }
                }
            }
        },
        "/api/v1/data/pembayaran": {
            "get": {
                "tags": [
                    "Data"
                ],
                "summary": "ambil data pembayaran",
                "description": "ambil data pembayaran",
                "operationId": "pembayaran",
                "responses": {
                    "200": {
                        "description": "sukses ambil data pembayaran"
                    },
                    "400": {
                        "description": "gagal mendapatkan data pembayaran"
                    }
                }
            }
        },
        "/api/v1/data/statuspembayaran": {
            "get": {
                "tags": [
                    "Data"
                ],
                "summary": "ambil data status pembayaran",
                "description": "ambil data status pembayaran",
                "operationId": "statuspembayaran",
                "responses": {
                    "200": {
                        "description": "sukses ambil data status pembayaran"
                    },
                    "400": {
                        "description": "gagal mendapatkan data status pembayaran"
                    }
                }
            }
        },
...
    }
}